+++ packages/health/lib/src/units.dart
@@ -15,6 +15,6 @@ enum HealthDataUnit {
   NO_UNIT,
   PERCENTAGE,
   SIEMENS,
-  UNKNOWN_UNIT
+  UNKNOWN_UNIT,
+  LITER,
 }
-